Moved function extraction back to mykutils; docstring
